+++ Scripts/Hearthstone Game Data API/Card Backs/Endpoint_Hearthstone_GameData_FetchOneCardBack.cs
@@ -15,35 +15,35 @@ namespace ZeShmouttsAssets.BlizzardAPI
 		/// </summary>
 		public static partial class HearthstoneGameData
 		{
-			internal const string apiPath_FetchOneCard = "/hearthstone/cards/{0}";
+			internal const string apiPath_FetchOneCardBack = "/hearthstone/cardbacks/{0}";
 
 			/// <summary>
-			/// Coroutine that retrieves a single card by ID.
+			/// Coroutine that retrieves a single card back by ID.
 			/// </summary>
-			/// <param name="cardId">The ID of the card.</param>
+			/// <param name="cardBackId">The ID of the card back.</param>
 			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
 			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
 			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
 			/// <param name="region">The region of the data to retrieve.</param>
 			/// <returns></returns>
-			public static IEnumerator FetchOneCard(int cardId, Action<Json_Hearthstone_Card> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			public static IEnumerator FetchOneCardBack(int cardBackId, Action<Json_Hearthstone_CardBack> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
 			{
-				string url = string.Concat(UrlDomain(region), string.Format(apiPath_FetchOneCard, cardId));
+				string url = string.Concat(UrlDomain(region), string.Format(apiPath_FetchOneCardBack, cardBackId));
 				yield return CustomRequest(url, null, action_Result, ifModifiedSince, action_LastModified);
 			}
 
 			/// <summary>
-			/// Coroutine that retrieves a single card by ID, as a raw JSON string.
+			/// Coroutine that retrieves a single card back by ID, as a raw JSON string.
 			/// </summary>
-			/// <param name="cardId">The ID of the card.</param>
+			/// <param name="cardBackId">The ID of the card back.</param>
 			/// <param name="action_Result">Action to execute with the raw JSON string.</param>
 			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
 			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
 			/// <param name="region">The region of the data to retrieve.</param>
 			/// <returns></returns>
-			public static IEnumerator FetchOneCardRaw(int cardId, Action<string> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			public static IEnumerator FetchOneCardBackRaw(int cardBackId, Action<string> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
 			{
-				string url = string.Concat(UrlDomain(region), string.Format(apiPath_FetchOneCard, cardId));
+				string url = string.Concat(UrlDomain(region), string.Format(apiPath_FetchOneCardBack, cardBackId));
 				yield return CustomRequest(url, null, action_Result, ifModifiedSince, action_LastModified);
 			}
 		}
@@ -56,31 +56,15 @@ namespace ZeShmouttsAssets.BlizzardAPI.JSON
 	/// JSON structure for Hearthstone, representing a single card.
 	/// </summary>
 	[Serializable]
-	public class Json_Hearthstone_Card : Object_JSON
+	public class Json_Hearthstone_CardBack : Object_JSON
 	{
 		// {{JSON_START}}
 		public int id;
-		public int collectible;
-		public string slug;
-		public int classId;
-		public int[] multiClassIds;
-		public int cardTypeId;
-		public int cardSetId;
-		public int rarityId;
-		public string artistName;
-		public int health;
-		public int attack;
-		public int manaCost;
-		public int armor;
-		public int durability;
-		public LocalizedString name;
+		public int sortCategory;
 		public LocalizedString text;
-		public LocalizedString image;
-		public LocalizedString imageGold;
-		public LocalizedString flavorText;
-		public string cropImage;
-		public int[] childIds;
-		public int[] keywordIds;
+		public LocalizedString name;
+		public string image;
+		public string slug;
 		// {{JSON_END}}
 	}
 }