+++ .github/CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project *somewhat* adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2020-06-07
+### Added
+ - Added support for the following Hearthstone Game Data endpoint category :
+   - Card Backs
+
+### Changed
+ - Slightly changed the Hearthstone card search parameters to use a new, auto-generated list of values
+
+### Bug fixes
+ - Fixed an incorrect summary in `FetchOneCard`
+ - Fixed a bug where the API Testing Tool would blank out and throw errors after a recompile
+
 ## [1.4.0] - 2020-06-08
 ### Added
  - Added support for the following WoW Classic Game Data endpoint category :